requirements.txt
textblob
newsapi-python
matplotlib
streamlit
python-dotenv 